Using secure version for inch-ci.org
Currently the status page of dry-rb.org is showing a warning:

   This website contains content that is not secure (such as images)

The warning is triggered by the "mixed content" the browser is requesting
for the inch-ci.org images, even thou inch-ci.org do the redirection
on their side.

With this change, the nagging warning will be gone since all the page
will be requesting only secure version (https).